+++ src/main/java/lotto/controller/LottoController.java
@@ -1,0 +1,5 @@
+package lotto.controller;
+
+public class LottoController {
+
+}
