+++ src/main/java/lotto/model/Product.java
