+++ src/main/java/lotto/model/Store.java
