+++ src/main/java/lotto/service/LottoService.java
@@ -1,0 +1,5 @@
+package lotto.service;
+
+public class LottoService {
+
+}
