+++ src/main/java/lotto/view/InputTable.java
@@ -1,0 +1,5 @@
+package lotto.view;
+
+public class InputTable {
+
+}
