+++ src/main/java/lotto/view/OutputTable.java
@@ -1,0 +1,5 @@
+package lotto.view;
+
+public class OutputTable {
+
+}
